Change max_tries/max_simultaneous from u32 to i32 in API types and config
Unify integer types so the API-facing types (AgentRecord, UpsertAgentRequest)
and config constants match the domain Agent struct (i32), eliminating all
as-i32/as-u32 boundary casts in routes/agents.rs, spawner.rs, and
run_spawners.rs.

+++ metis-server/src/routes/agents.rs
@@ -158,8 +158,8 @@ fn normalize_and_build_agent(
     let agent = Agent::new(
         name,
         prompt_path,
-        payload.max_tries as i32,
-        payload.max_simultaneous as i32,
+        payload.max_tries,
+        payload.max_simultaneous,
         payload.is_assignment_agent,
     );
 
@@ -179,8 +179,8 @@ fn agent_to_record(agent: Agent, prompt: String) -> AgentRecord {
         agent.name,
         prompt,
         agent.prompt_path,
-        agent.max_tries as u32,
-        agent.max_simultaneous as u32,
+        agent.max_tries,
+        agent.max_simultaneous,
         agent.is_assignment_agent,
     )
 }
